refactor: move more functions into layer architecture
