+++ backend/src/handlers/write.js
@@ -2,7 +2,7 @@ const addCommentToVideo = require("../application/add-comment-to-video");
 const addTag = require("../persistence/add-tag");
 const addViewToVideo = require("../application/add-view-to-video");
 const deleteTag = require("../persistence/delete-tag");
-const deleteVideo = require("../lib/delete-video");
+const deleteVideo = require("../persistence/delete-video");
 const ResponseFactory = require("../lib/factories/ResponseFactory");
 const getAuthToken = require("../utility/get-auth-token");
 const isUserAnAdmin = require("../lib/is-user-an-admin");
